--- news-JS/src/components/controller/appLoader.js
@@ -1,11 +1,0 @@
-import Loader from './loader';
-
-class AppLoader extends Loader {
-    constructor() {
-        super('https://newsapi.org/v2/', {
-            apiKey: '64264965e7994d6f8aaac6b1c6e43c5d', 
-        });
-    }
-}
-
-export default AppLoader;
